move getContainersURL to its own file
Issue: AAH-2518
