+++ frontend/hub/common/api/hub-api-utils.tsx
@@ -308,19 +308,3 @@ export function getRepoURL(distribution_base_path: string, view_published = fals
 
   return `${host}${base}/content/${distribution_base_path}/`;
 }
-
-// returns the server name for (protocol-less) container urls
-// url/image, url/image:tag, url/image@digest (including sha256: prefix)
-export function getContainersURL({
-  name,
-  tag,
-  digest,
-}: {
-  name: string;
-  tag?: string;
-  digest?: string;
-}) {
-  const host = window.location.host;
-
-  return `${host}/${name}${tag ? `:${tag}` : ''}${digest && !tag ? `@${digest}` : ''}`;
-}
